- Ürüne özel operasyon kartlarına kalite planı eklendi.
- Ürüne özel operasyon kalite planı Get sorunu giderildi.
- Login sayfası tasarımı düzenlendi. İyileştirmeler yapılacak.

+++ TsiErp.Entities/Entities/ProductionManagement/ProductOperationQualtityPlan/Dtos/CreateProductOperationQualityPlansDto.cs
@@ -75,6 +75,6 @@ namespace TsiErp.Entities.Entities.ProductionManagement.ProductOperationQualtity
         /// <summary>
         /// Resimdeki Ölçü Numarası
         /// </summary>
-        public int MeasureNumberInPicture { get; set; }
+        public decimal MeasureNumberInPicture { get; set; }
     }
 }
